+++ CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/iterativelyhq/itly-sdk/compare/v1.1.2...v1.2.0) (2020-08-22)
+
+
+### Bug Fixes
+
+* changed mParticle instance initialization check ([d516c8a](https://github.com/iterativelyhq/itly-sdk/commit/d516c8a292ca6ad54185409daa8ffcfb339362f1))
+* don't re-initialize mParticle instance if it initialized already ([8fb9af5](https://github.com/iterativelyhq/itly-sdk/commit/8fb9af5f25b17044da1531975b30686a653c9973))
+* exposed EventMetadata type in sdk ([7713f07](https://github.com/iterativelyhq/itly-sdk/commit/7713f07bef0d5ee8a1bd3856c2b2e84c0f12a205))
+* provided env specific default options for options.validation ([947d44a](https://github.com/iterativelyhq/itly-sdk/commit/947d44afe193416052e828999bc901eb7e7ec809))
+* refreshed plugin-mparticle's yarn.lock ([112710a](https://github.com/iterativelyhq/itly-sdk/commit/112710a80d4478dc7d999cdcf11e37ccd4acaf82))
+* remove mParticle mention from sdk-base Event type ([7bbd330](https://github.com/iterativelyhq/itly-sdk/commit/7bbd330be8d79c641bea1941bca97f0177a70cb3))
+* removed itly from MparticleOptions ([39ec98b](https://github.com/iterativelyhq/itly-sdk/commit/39ec98bac3cc25fc211425a746d0e3a5c2b9d181))
+* use Mparticle.getInstance to return initialized mParticle instance ([225f543](https://github.com/iterativelyhq/itly-sdk/commit/225f5439afa8b16380b7c444e9a28f7e3b5a5201))
+
+
+### Features
+
+* added a draft of mparticle browser plugin ([ed4ff71](https://github.com/iterativelyhq/itly-sdk/commit/ed4ff71437a51cbcbdb6cfe54960c2c29fae6383))
+
+
+
+
+
 ## [1.1.2](https://github.com/iterativelyhq/itly-sdk/compare/v1.1.1...v1.1.2) (2020-08-05)